+++ src/commands/sync.tsx
@@ -10,13 +10,6 @@ function Sync({cli, input}: CommandProps) {
 export const syncConfig = {
 	description: 'Sync your local changes with the remote server',
 	usage: 'sync',
-	validateProps({cli, input}: CommandProps) {
-		console.log({cli, input});
-		return {
-			valid: true,
-			errors: [],
-		};
-	},
 };
 
 export default Sync;
