+++ src/hooks/useGit.ts
@@ -1,0 +1,12 @@
+import {simpleGit, SimpleGit, SimpleGitOptions} from 'simple-git';
+
+const options: Partial<SimpleGitOptions> = {
+	baseDir: process.cwd(),
+	binary: 'git',
+	maxConcurrentProcesses: 6,
+	trimmed: false,
+};
+
+export const useGit = (): SimpleGit => {
+	return simpleGit(options);
+};
